+++ packages/nx-plugin/src/utils/versions/dev-dependencies.ts
@@ -111,15 +111,15 @@ const getDevDependencies = (
         ? V19_X_VITE
         : NX_X_LATEST_VITE,
     vitest:
-      escapedNxVersion && lt(escapedNxVersion, '20.5.0')
+      escapedNxVersion && lt(escapedNxVersion, '22.3.0')
         ? V19_X_VITEST
         : NX_X_LATEST_VITEST,
     '@vitest/coverage-v8':
-      escapedNxVersion && lt(escapedNxVersion, '20.5.0')
+      escapedNxVersion && lt(escapedNxVersion, '22.3.0')
         ? V19_X_VITEST
         : NX_X_LATEST_VITEST,
     '@vitest/ui':
-      escapedNxVersion && lt(escapedNxVersion, '20.5.0')
+      escapedNxVersion && lt(escapedNxVersion, '22.3.0')
         ? V19_X_VITEST
         : NX_X_LATEST_VITEST,
   };